refactor: split files exceeding line caps
cell_entity.rs (701 -> mod.rs 456 + tests.rs 246): hard cap exceeded.
Production code stays in mod.rs; the test module moves to tests.rs as
a sibling submodule.

cell_dispatch.rs (523 -> 420): handle_teleport_player extracted to a
new world_entry/teleport.rs (123 lines).

messages.rs (566 -> messages/mod.rs 21 + base_to_cell.rs 150 +
cell_to_base.rs 296 + data.rs 57 + tests.rs 66): natural seam between
the two enum directions, plus the shared `MailOp` / `NpcAoIData` /
`SavedMission` structs in `data`.

ring_transport_runtime.rs (744, over hard cap) moves into the
ring_transport/ directory and splits into:
- runtime.rs (252) — public entry points (handle_interact,
  handle_select_destination, handle_region_trigger,
  run_tick_with_engine, kick_off_warmup,
  advance_destination_after_warmup).
- dispatch.rs (207) — Effect dispatcher, mark_player_loaded,
  try_advance_after_load, same_world_teleport.
- wire_helpers.rs (145) — onSequence/onVisible/onStateFieldUpdate
  payload helpers + BSF_MOVEMENT_LOCK constant.
- tests.rs (210) — integration-style tests against a real SpaceManager.

Call sites updated from `crate::cell::ring_transport_runtime::xxx` to
`crate::cell::ring_transport::xxx`.

transporter.rs (664 -> 682, in 500-700 range): split skipped. The
suggested seam (manager + Effect ~80 lines vs FSM ~600 lines) produces
uneven files with no real readability win — kept as one cohesive FSM
module per the CLAUDE.md "skip the split if the only seams produce
uneven files" rule, with a doc comment recording the decision.
