--- crates/services/src/cell/messages.rs
@@ -1,571 +1,0 @@
-//! Message types for Base↔Cell inter-service communication.
-//!
-//! In the C++ architecture, BaseApp and CellApp communicated over Mercury UDP.
-//! In our single-process Rust server, they use `tokio::mpsc` channels with these
-//! typed message enums.
-
-/// Mail operation types forwarded from CellService to BaseApp for DB execution.
-#[derive(Debug)]
-pub enum MailOp {
-    /// Request mail headers (inbox or archive).
-    RequestHeaders { b_archive: u8 },
-    /// Request a specific mail body.
-    RequestBody { mail_id: i32 },
-    /// Delete a mail message.
-    Delete { mail_id: i32 },
-    /// Archive a mail message.
-    Archive { mail_id: i32 },
-}
-
-/// NPC-specific data included in AoI enter events.
-///
-/// Carries template-driven values that the client needs for correct rendering
-/// and interaction display. Only populated for NPC entities (not players).
-///
-/// Mirrors the full `createOnClient()` cascade from the Python scripts:
-/// `SGWSpawnableEntity.createOnClient()` → `SGWBeing.createOnClient()`.
-#[derive(Debug, Clone, Default)]
-pub struct NpcAoIData {
-    /// Localized name string ID from `entity_templates.name_id`.
-    pub name_id: Option<i32>,
-    /// Faction ID (0=neutral, 1=Tau'ri, 3=SGC, 10=hostile).
-    pub faction: u8,
-    /// Alignment ID.
-    pub alignment: u8,
-    /// Entity flags from `entity_templates.flags`.
-    pub entity_flags: u64,
-    /// Interaction type flags (UINT64 bitmask for cursor/interaction UI).
-    pub interaction_type: i64,
-    /// Speaker ID for `onEntityProperty(GENERICPROPERTY_DatabaseId, speakerId)`.
-    pub speaker_id: Option<i32>,
-    /// Kismet event set ID for `onKismetEventSetUpdate`.
-    pub event_set_id: Option<i32>,
-    /// Static mesh name (for `onStaticMeshNameUpdate` — non-humanoid entities).
-    pub static_mesh: Option<String>,
-    /// Body set name (for `BeingAppearance` — humanoid entities, or `onStaticMeshNameUpdate`).
-    pub body_set: Option<String>,
-    /// Body components (for `BeingAppearance` — humanoid entities with body parts).
-    pub components: Vec<String>,
-}
-
-/// A saved mission loaded from the database for re-login.
-#[derive(Debug, Clone)]
-pub struct SavedMission {
-    pub mission_id: i32,
-    pub status: i8,
-    pub current_step_id: Option<i32>,
-    pub completed_step_ids: Vec<i32>,
-    pub completed_objective_ids: Vec<i32>,
-    pub active_objective_ids: Vec<i32>,
-    pub failed_objective_ids: Vec<i32>,
-}
-
-/// Messages sent from BaseApp to CellApp.
-// Cannot derive Debug because oneshot::Sender doesn't implement Debug.
-// Manual impl would be possible but not worth the boilerplate.
-pub enum BaseToCellMsg {
-    /// Create a cell entity in the named world at the given position/rotation.
-    /// The `reply_tx` oneshot returns the resolved `space_id` so the caller
-    /// can `.await` it before building the world-entry wire packet.
-    CreateEntity {
-        entity_id: u32,
-        world_name: String,
-        position: [f32; 3],
-        rotation: [f32; 3],
-        reply_tx: tokio::sync::oneshot::Sender<u32>,
-    },
-
-    /// Destroy a cell entity (player left, entity despawned).
-    DestroyEntity {
-        entity_id: u32,
-    },
-
-    /// Mark an entity as having a client controller (player).
-    /// Sent after world entry packets are delivered to the client.
-    ConnectEntity {
-        entity_id: u32,
-    },
-
-    /// Remove client controller from an entity (player disconnected).
-    DisconnectEntity {
-        entity_id: u32,
-    },
-
-    /// Client position/movement update forwarded from `avatarUpdateExplicit`.
-    EntityMove {
-        entity_id: u32,
-        position: [f32; 3],
-        direction: [i8; 3],
-        velocity: [f32; 3],
-    },
-
-    /// Client→server cell entity method call forwarded from BaseApp.
-    ///
-    /// `method_index` is the flattened EXPOSED CellMethod index for the
-    /// SGWPlayer entity type (0 = setTargetID, 1 = setMovementType, etc.).
-    /// `args` contains the raw method arguments (after entity_id extraction).
-    CellMethodCall {
-        entity_id: u32,
-        method_index: u16,
-        args: Vec<u8>,
-    },
-
-    /// Chat message from a player, forwarded from BaseApp for spatial distribution.
-    ///
-    /// The CellService broadcasts to witnesses based on channel type:
-    /// say/emote (nearby), yell (wider range).
-    ChatMessage {
-        entity_id: u32,
-        speaker_name: String,
-        speaker_flags: u8,
-        channel: u8,
-        text: String,
-    },
-
-    /// Initialize player state after world entry (missions, etc.).
-    /// Sent after ConnectEntity so the CellService can populate per-player data.
-    InitPlayerState {
-        entity_id: u32,
-        player_id: i32,
-        world_name: String,
-        /// Saved missions loaded from DB, to be restored before content engine fires.
-        saved_missions: Vec<SavedMission>,
-        /// Player's known ability IDs (from sgw_player.abilities column).
-        abilities: Vec<i32>,
-        /// Active bandolier slot from `sgw_player.bandolier_slot` (0-based).
-        active_bandolier_slot: i32,
-        /// Bandolier slot contents loaded from `sgw_inventory` + `resources.items`.
-        bandolier_items: Vec<(i32, cimmeria_entity::cell_entity::BandolierItem)>,
-    },
-
-    /// Update one bandolier slot after a runtime item grant.
-    ///
-    /// BaseApp persists inventory changes and sends the client inventory update,
-    /// while CellApp owns combat state. This keeps the cell-side weapon cache in
-    /// sync without waiting for relog/world entry.
-    UpdateBandolierItem {
-        entity_id: u32,
-        slot_id: i32,
-        item: cimmeria_entity::cell_entity::BandolierItem,
-        make_active: bool,
-    },
-
-    /// Replace the whole cell-side bandolier cache after inventory move/remove.
-    SyncBandolierItems {
-        entity_id: u32,
-        active_bandolier_slot: i32,
-        bandolier_items: Vec<(i32, cimmeria_entity::cell_entity::BandolierItem)>,
-    },
-
-    /// Inventory move committed in BaseApp after DB validation.
-    ///
-    /// CellApp uses the source/target transition to fire item equip/unequip
-    /// event abilities only after the move is known to have persisted.
-    InventoryItemMoveApplied {
-        entity_id: u32,
-        item_id: i32,
-        source_container_id: i32,
-        target_container_id: i32,
-        swapped_item_id: Option<i32>,
-    },
-
-    /// Inventory item instance was fully removed after DB validation.
-    InventoryItemRemoved {
-        entity_id: u32,
-        item_id: i32,
-        source_container_id: i32,
-    },
-
-    /// Inventory item was granted and persisted in BaseApp.
-    InventoryItemGranted {
-        entity_id: u32,
-        item_id: i32,
-        container_id: i32,
-        slot_id: i32,
-        quantity: i32,
-    },
-
-    /// Inventory item was consumed atomically by base (in response to
-    /// `CellToBaseMsg::UseInventoryItem`). The cell should fire the
-    /// `OnItemUse` content event with `type_id` (item design id). Only sent
-    /// after the consumption tx commits — if the player didn't own that
-    /// instance or the tx failed, this message is not emitted.
-    ItemUsed {
-        entity_id: u32,
-        type_id: i32,
-        target_id: i32,
-    },
-
-    /// Reload the content engine from the database (triggered by admin API / Content Editor).
-    ReloadContentEngine,
-
-    /// Minigame result callback (forwarded from BaseApp after minigame server reports).
-    MinigameResult {
-        entity_id: u32,
-        result_code: u8,
-        on_victory_chains: Vec<i64>,
-    },
-}
-
-/// Messages sent from CellApp to BaseApp.
-#[derive(Debug)]
-pub enum CellToBaseMsg {
-    /// Notification that a space exists (sent at startup and on dynamic creation).
-    SpaceData {
-        space_id: u32,
-        world_name: String,
-    },
-
-    /// Response to `CreateEntity` — entity placed in a space.
-    EntityCreated {
-        entity_id: u32,
-        space_id: u32,
-        position: [f32; 3],
-    },
-
-    /// An entity has moved (for relaying volatile updates to a specific witness via BaseApp).
-    EntityMoved {
-        witness_id: u32,
-        entity_id: u32,
-        space_id: u32,
-        position: [f32; 3],
-        direction: [f32; 3],
-        velocity: [f32; 3],
-    },
-
-    /// A new entity entered a witness's Area of Interest.
-    EnteredAoI {
-        witness_id: u32,
-        entity_id: u32,
-        space_id: u32,
-        class_id: u8,
-        position: [f32; 3],
-        direction: [f32; 3],
-        /// Entity level (for `onLevelUpdate`). Defaults to 1.
-        level: u32,
-        /// NPC-specific data (faction, alignment, flags, name). None for players.
-        npc_data: Option<NpcAoIData>,
-    },
-
-    /// An entity left a witness's Area of Interest.
-    LeftAoI {
-        witness_id: u32,
-        entity_id: u32,
-    },
-
-    /// Send a server→client entity method call to the entity's client.
-    ///
-    /// The BaseApp looks up the entity's client address and sends the method
-    /// call using `append_entity_method` encoding.
-    EntityMethodCall {
-        entity_id: u32,
-        method_index: u16,
-        args: Vec<u8>,
-    },
-
-    /// Request mail headers for a player entity (forwarded to BaseApp for DB query).
-    MailRequest {
-        entity_id: u32,
-        player_id: i32,
-        /// Mail operation type.
-        op: MailOp,
-    },
-
-    /// Request a gate travel (world transition) for a player entity.
-    ///
-    /// The CellService has already validated the stargate address and removed
-    /// the entity from the old space. BaseApp must send RESET_ENTITIES to the
-    /// client, then re-create the entity in the new world via the standard
-    /// World entry flow (teardown -> create player -> enter world).
-    GateTravel {
-        entity_id: u32,
-        target_world_name: String,
-        position: [f32; 3],
-        rotation: [f32; 3],
-    },
-
-    /// Persist a mission state change to the database.
-    ///
-    /// Uses `INSERT ... ON CONFLICT DO UPDATE` on `sgw_mission`.
-    MissionUpdate {
-        player_id: i32,
-        mission_id: i32,
-        status: i8,
-        current_step_id: Option<i32>,
-        completed_step_ids: Vec<i32>,
-        completed_objective_ids: Vec<i32>,
-        active_objective_ids: Vec<i32>,
-        failed_objective_ids: Vec<i32>,
-    },
-
-    /// Grant XP to a player entity (from a mob kill).
-    ///
-    /// The CellService computes the XP amount from the mob's level and sends
-    /// this to BaseApp, which updates the player's XP/level and sends client
-    /// notifications.
-    GrantXP {
-        entity_id: u32,
-        xp_amount: u64,
-    },
-
-    /// Grant an item to a player and persist to `sgw_inventory`.
-    GrantItem {
-        entity_id: u32,
-        player_id: i32,
-        item_id: i32,
-        container_id: i32,
-        count: i32,
-    },
-
-    /// Open a vendor store for a player using the vendor template lists.
-    OpenVendorStore {
-        entity_id: u32,
-        player_id: i32,
-        vendor_entity_id: i32,
-        vendor_template_id: Option<i32>,
-    },
-
-    /// Purchase items from the currently-open vendor store.
-    PurchaseVendorItems {
-        entity_id: u32,
-        player_id: i32,
-        vendor_entity_id: i32,
-        vendor_template_id: i32,
-        items: Vec<(i32, i32)>,
-    },
-
-    /// Sell owned inventory items to the currently-open vendor store.
-    SellVendorItems {
-        entity_id: u32,
-        player_id: i32,
-        vendor_entity_id: i32,
-        vendor_template_id: i32,
-        items: Vec<(i32, i32)>,
-    },
-
-    /// Buy back recently-sold inventory items from the currently-open vendor store.
-    BuybackVendorItems {
-        entity_id: u32,
-        player_id: i32,
-        vendor_entity_id: i32,
-        vendor_template_id: i32,
-        items: Vec<(i32, i32)>,
-    },
-
-    /// Request a full inventory refresh from BaseApp.
-    ListInventoryItems { entity_id: u32, player_id: i32 },
-
-    /// Move an inventory item instance to another bag/slot.
-    MoveInventoryItem {
-        entity_id: u32,
-        player_id: i32,
-        item_id: i32,
-        target_container_id: i32,
-        target_slot_id: i32,
-        quantity: i32,
-    },
-
-    /// Remove quantity from an inventory item instance.
-    RemoveInventoryItem {
-        entity_id: u32,
-        player_id: i32,
-        item_id: i32,
-        quantity: i32,
-    },
-
-    /// Consume one charge/stack of an inventory item instance, then fire the
-    /// content-engine `OnItemUse` event back to the cell with the resolved
-    /// `type_id` (item design id). Mission progression that depends on item
-    /// use is gated on the consumption committing successfully — if the row
-    /// can't be found or the tx fails, no event fires and the mission does
-    /// not advance.
-    ///
-    /// `item_id` is the inventory instance id from the wire (`useItem` arg
-    /// per `SGWInventoryManager.def` — "Player inventory id"). The type_id is
-    /// looked up server-side and returned via `BaseToCellMsg::ItemUsed`.
-    UseInventoryItem {
-        entity_id: u32,
-        player_id: i32,
-        item_id: i32,
-        target_id: i32,
-    },
-
-    /// Repair an owned inventory item instance by a durability ratio.
-    RepairInventoryItem {
-        entity_id: u32,
-        player_id: i32,
-        item_id: i32,
-        repair_ratio: f32,
-    },
-
-    /// Fully repair owned inventory item instances.
-    RepairInventoryItems {
-        entity_id: u32,
-        player_id: i32,
-        item_ids: Vec<i32>,
-        vendor_template_id: Option<i32>,
-    },
-
-    /// Fully recharge owned inventory item instances.
-    RechargeInventoryItems {
-        entity_id: u32,
-        player_id: i32,
-        item_ids: Vec<i32>,
-        vendor_template_id: Option<i32>,
-    },
-
-    /// Persist the player's active bandolier slot.
-    ActiveSlotUpdate { player_id: i32, slot_id: i32 },
-
-    /// Persist a single bandolier slot's per-slot ammo state.
-    ///
-    /// Sent by the cell on the batched cadence (reload completion, slot swap,
-    /// `requestAmmoChange`, logout) — see Stage B/C/D in
-    /// docs/gameplay/weapon-ammo-reload.md (TBD). `player_id` here matches the
-    /// DB `character_id`, mirroring the field naming used by `ActiveSlotUpdate`.
-    ///
-    /// `expected_item_id` guards against TOCTOU: if the slot's item changes
-    /// between the cell sending this message and the base writing the row,
-    /// the SQL `WHERE type_id = $expected_item_id` clause skips the write
-    /// rather than scribbling stale ammo onto the new weapon.
-    BandolierAmmoUpdate {
-        player_id: i32,
-        slot_id: i32,
-        expected_item_id: i32,
-        current_ammo: i32,
-        cur_ammo_type: i32,
-    },
-
-    /// Grant cash (naquadah) to a player and persist to the database.
-    GrantCash {
-        entity_id: u32,
-        player_id: i32,
-        amount: i32,
-    },
-
-    /// Respawn reload: send onClientMapLoad to trigger a loading screen, then
-    /// set up pending_client_ready so the next onClientReady triggers a fresh
-    /// mapLoaded sequence with cleared state. This is the only reliable way to
-    /// reset ragdoll/death state on the client.
-    RespawnReload {
-        entity_id: u32,
-        world_name: String,
-        spawn_pos: [f32; 3],
-    },
-
-    /// Authoritative same-world teleport. The cell has already updated its
-    /// spatial state via `update_entity_position`; the base must now push the
-    /// new position to the player's own client. Engine-level snap, no loading
-    /// screen, no world reload — used by ring transporters and any other
-    /// in-world short-hop teleport.
-    ///
-    /// `space_id` is the cell's authoritative space — the base trusts this
-    /// over its connection-cached `world_name`, which can lag during world
-    /// transitions.
-    ///
-    /// Other witnesses receive their position update through the next AoI
-    /// tick's `EntityMoved` broadcast — no extra fan-out is needed here.
-    TeleportPlayer {
-        entity_id: u32,
-        space_id: u32,
-        position: [f32; 3],
-    },
-
-    /// Send a ghost entity method call to a specific witness player.
-    ///
-    /// Used for broadcasting property updates (InteractionType, SetVisible, etc.)
-    /// to players who have the entity in their AoI. The `entity_id` is the ghost
-    /// entity the method is called on; `witness_id` is the player to send to.
-    WitnessEntityMethod {
-        witness_id: u32,
-        entity_id: u32,
-        method_index: u16,
-        args: Vec<u8>,
-    },
-
-    /// Start a minigame session for a player (Cell → Base).
-    StartMinigame {
-        entity_id: u32,
-        player_id: i32,
-        game_name: String,
-        difficulty: u32,
-        on_victory_chains: Vec<i64>,
-    },
-
-    /// Minigame result callback (minigame server → Cell via Base).
-    MinigameResult {
-        entity_id: u32,
-        result_code: u8,
-        on_victory_chains: Vec<i64>,
-    },
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn saved_mission_debug_prints() {
-        let m = SavedMission {
-            mission_id: 622,
-            status: 1,
-            current_step_id: Some(700),
-            completed_step_ids: vec![],
-            completed_objective_ids: vec![],
-            active_objective_ids: vec![800, 801],
-            failed_objective_ids: vec![],
-        };
-        let debug = format!("{:?}", m);
-        assert!(debug.contains("622"));
-        assert!(debug.contains("800"));
-    }
-
-    #[test]
-    fn saved_mission_clone() {
-        let m = SavedMission {
-            mission_id: 622,
-            status: 1,
-            current_step_id: Some(700),
-            completed_step_ids: vec![699],
-            completed_objective_ids: vec![801],
-            active_objective_ids: vec![800],
-            failed_objective_ids: vec![],
-        };
-        let cloned = m.clone();
-        assert_eq!(cloned.mission_id, 622);
-        assert_eq!(cloned.status, 1);
-        assert_eq!(cloned.current_step_id, Some(700));
-        assert_eq!(cloned.completed_step_ids, vec![699]);
-        assert_eq!(cloned.completed_objective_ids, vec![801]);
-        assert_eq!(cloned.active_objective_ids, vec![800]);
-        assert!(cloned.failed_objective_ids.is_empty());
-    }
-
-    #[test]
-    fn saved_mission_completed_status() {
-        let m = SavedMission {
-            mission_id: 622,
-            status: 2, // MISSION_COMPLETED
-            current_step_id: None,
-            completed_step_ids: vec![700],
-            completed_objective_ids: vec![800, 801],
-            active_objective_ids: vec![],
-            failed_objective_ids: vec![],
-        };
-        assert_eq!(m.status, 2);
-        assert!(m.current_step_id.is_none());
-        assert_eq!(m.completed_step_ids.len(), 1);
-        assert_eq!(m.completed_objective_ids.len(), 2);
-    }
-
-    #[test]
-    fn npc_aoi_data_default() {
-        let data = NpcAoIData::default();
-        assert_eq!(data.faction, 0);
-        assert_eq!(data.alignment, 0);
-        assert_eq!(data.entity_flags, 0);
-        assert_eq!(data.interaction_type, 0);
-        assert!(data.name_id.is_none());
-        assert!(data.components.is_empty());
-    }
-}
