+++ crates/services/src/cell/mod.rs
@@ -17,7 +17,6 @@ pub mod mail;
 pub mod messages;
 pub mod missions;
 pub mod ring_transport;
-pub mod ring_transport_runtime;
 mod service;
 pub mod space_manager;
 pub mod spawner;
